Rough reimplementation of current algorithm.
Reimplemented using raw JGit rather than Grgit for performance.

There's a commented out section of code that should implement
the new piece for _claiming_ versions. Right now this causes
the tests to fail. Need to investigate if this is because the
test is incorrect given the new algorithm, or if the implementation
is wrong.