Added impl of full inventory
Inventory will be the base of information needed from Git to make a
decision on the version. All other information will be external (i.e.
scope/stage selection).

This impl needs to be tested to ensure the correct values are picked up.
In particular, I'm not comfortable with the parallel version
implementation.