Documentation improvements
Broke out "how" documentation into a separate doc and made sure all of
the sentences finished. Also addressed a few specific pieces of
feedback.

More still needs to be done to describe all of the behaviors with
parallel development and the general inference algorithm.

This fixes #62 and fixes #58 and fixes #55.