Ignore reckon properties in composite builds
Composite builds that include multiple builds that use reckon can run
into unexpected behavior, because any reckon.* props passed in on the
command line are used by all builds. This can cause it to try to
increment a version incorrectly.

The new behavior makes it so that only the main build within a composite
will actually respect the gradle properties.

This does imply that you can't really release a version of an included
build, but the expectation is that that's not what anyone would intend
to do. That could be proven wrong, but it would require something more
involved to support.

This fixes #74.