Support reckon.stage for snapshot strategy
The reckon.snapshot property is a little confusing to set the value of,
since true means snapshot and final is false. It seems a little
backwards. The reckon.stage approach is a little more natural,
explicitly passing in 'snapshot' or 'final'.

This will deprecate use of reckon.snapshot, to be removed in 1.0.0.

This fixes #71.