Use abbreviated hash in build metedata
Since the commit ID is only used for metadata on insignificant versions,
there's no reason to be so concerned about uniqueness that we use a full
hash. It will be easier for users to deal with if its shorter. Using the
ObjectReader method in this change, does ensure uniqueness within the
current repo, which is all we really need anyway.

This fixes #86.