Make uncomitted versions unique
Instead of using the hash, if present, and the string 'uncomitted', use
a UTC timestamp. Unique versions allow you to go back to a specific
build more easily.

This may make it harder to distinguish between commit builds and
uncommitted builds, so I could see us adding some identifier back, but
this keeps the versions shorter for now.

This fixes #90.