Don't treat snapshot as a significant version
We previously failed the build if you tried to explicitly build a
snapshot on a dirty repo. This is because we treated it the same as for
the stage-style approaches where input always indicates a significant
version. Now we do the validation after we have the version and can just
explicitly check for it being a significant version.

This fixes #85.