Fail on invalid increment in Scope.infer
When the base normal and base version aren't within 1 major/minor/patch
increment, there's something weird going on in the history, because the
versions aren't contiguous. This may be something we need to support for
parallel development at some point, but right now this makes it more
explicit and allows specific use cases to be raised.