Fixed rebuild behavior with parallel branches
In some cases the inferred target normal would be different from the
target normal of the final reckoned version. This can cause an issue if
you're trying to release an older version and the inferred target normal
is a version you already released.

The code now verifies the normal version of the final reckoned version,
rather than relying on the potentially incorrect target normal.

This fixes #108.