Allow pre-release versions on tagged final
If the current version is a final version, we previously did not allow
you to do anything besides rebuild as the current version or increment
to another final version.

While this isn't a recommended option, some people want to be able to
release pre-releases from the same commit that was already released.

There isn't anything critical in reckon depending on preventing this
kind of release, so it seems OK to allow.