minor: Allow setting assumed scope of parallel branches
This new approach replaces the one introduced in [1]. You can now
specify a parallelBranchScope in the extension, which defaults to
PATCH. This implies the scope that a paralle branch "claims" or
"owns".

e.g. if you use branches like maintenance/1.2.x, set it to MINOR. If
you set it to maintenance/1.x, set it to MAJOR.

In cases where a parallel branch is found that collides with the
normal that would have otherwise been inferred, we will try another
increment. In the past we would just re-increment with the scope that
was provided. In the new model we'll increment with the greater of the
user-provided scope and the parallelBranchScope.

[1] 9152df59dd1ff0fc23f879c2ba48f55e2b27a412