Only use parallel version avoidance for significants
Only if reckoning a significant version should we try to avoid
version numbers tagged on parallel branches. This is in line with our
existing documented axioms (which focused on two TAGGED parallel
versions).

What this helps with is feature branches where you aren't intending to
make a release, and there's more benefit to having the version
logically related to what the history of the branch is, rather than
assuming the feature branch will eventually conflict and try to
eagerly avoid the existing parallel version.