Don't soft-fail in settings plugin
The settings plugin was intended to avoid the issue of another plugin
calling version.toString() to early. As such, it doesn't need the
workaround added to the original plugin allowing for a soft-fail.

As noted in #194, the soft-fail can be detrimental in normal cases, as
ordinary failures you want to be informed of get covered up.

Fixes #194