Only soft fail on config errrors
While we don't want to soft-fail at all in the settings plugin, we still
want to provide this as a workaround for misbehaving plugins that might
interfere with your version before it's configured.

However, as #194 made clear, we want to be more careful about which
exceptions we catch. We'd really like to catch cases of failing to build
the reckoner configuration. We do need to inspect the throwable cause
chain though given the layers Gradle adds on top.

Fixes #194