minor: Allow commit messages to bump to 1.0.0
As reported in #204, I had unfortunately forced the preference to not
allow commit messages to bump to v1 on all users of commit message
scope calculators. In theory, I had meant to allow custom parsing, but
the custom parsing still enforced the pre-1.0 special casing.

The new approach implements a new ofCommitMessageParser() mechanism
that has the parsing function accept both the commit message and a
boolean indicating whether the project is pre-v1.0.0. This allows the
user full control of the approach they want to use.

To simplify adoption of this new capability the default commit message
parser was enhanced to allow "major!: subject" prefixes that bypass
the pre-1.0.0 check, allowing a more CD style bump to 1.0.0.

Fixes #204