patch: Fix configuration caching support
Gradle doesn't support external processes at configuration time if
configuration cache is turned on.

JGit has an unfortunate behvior [1] (at least in the context of
Gradle), where it uses calls to the external git executable to
determine the location of system configuration files. This workaround
allows blocks the use of system configuration as part of Reckon's
calculation, but restores it for use by any later JGit-based logic.

Fixes #202

[1] https://github.com/eclipse-jgit/jgit/issues/24